scanning/db/requirements.txt
mysql-connector-python~=8.3.0
psycopg2~=2.9.9